Changes requested done
